Successfully implemented the MVP
It's quite messy but I was able to get the result that I want in the
command line. The user will enter a stock symbol, enter its Buy Below
and Target Price and hit Enter. Then, Sammy will make a request to the
API that I use to determine the current price of stock chosen. This is
messy but I'm on the right path :)