Refresh and Editing enhancement implemented to Sammy
Did a bit of enhancement in the application.

(1) Added a 'Refresh' feature to re-request to the API. This feature is
useful when the initial API request failed so no need to re-run the app
but it needs refinement.

(2) User can now edit the cells in the table under the 'Remarks' column
but the application doesn't save it upon exit.